requirements.txt
transformers 
pdfplumber 
pandas 
torch 
sentence-transformers
streamlit
langchain
langchain-community
bcrypt
llama-cpp-python